requirements/base.txt
Django==1.5.1
Jinja2==2.6
Pygments==1.6
South==0.7.6
argparse==1.2.1
bpython==0.12
distribute==0.6.31
django-appconf==0.5
django-braces==1.0.0
django-compressor==1.2
django-forms-bootstrap==2.0.3.post1
django-model-utils==1.1.0
django-mptt==0.5.4
django-oembed==0.1.3
django-pagination==1.0.7
django-reversion==1.7
django-taggit==0.9.3
django-tastypie==0.9.11
docutils==0.10
ipython==0.13.1
logutils==0.3.3
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==1.5
python-memcached==1.48
sorl-thumbnail==11.12
wsgiref==0.1.2
textile==2.1.5
Pillow==2.0.0
